Стало нормальным api, без фронта

+++ src/main/java/com/example/finalfinalback3/Service/UserService.java
@@ -1,15 +1,12 @@
 package com.example.finalfinalback3.Service;
 
-import com.example.finalfinalback3.DTO.UserAuthDTO;
 import com.example.finalfinalback3.Entity.UserEntity;
-import com.example.finalfinalback3.Exceptions.PasswordsNotSameException;
-import com.example.finalfinalback3.Exceptions.UserAlreadyExistsException;
-import com.example.finalfinalback3.Exceptions.UserNotFoundException;
-import com.example.finalfinalback3.Model.User;
+import com.example.finalfinalback3.Exceptions.DataNotFoundException;
 import com.example.finalfinalback3.Repository.UserRepository;
-import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 @Service
 public class UserService {
@@ -18,19 +15,26 @@ public class UserService {
     private UserRepository userRepo;
 
     //Спросить насчёт использования Optional
-    public User getSingleUser(Integer id) throws UserNotFoundException{
+    public UserEntity getSingleUser(Integer id) throws DataNotFoundException {
         if (userRepo.findById(id).get() == null) {
-            throw new UserNotFoundException("Пользователя с таким id не существует!");
+            throw new DataNotFoundException("Пользователя с таким id не существует!");
         }
-        return User.toModel(userRepo.findById(id).get());
+        return userRepo.findById(id).get();
     }
 
-    public Integer deleteUser(Integer id){
+    public List<UserEntity> showAll() throws DataNotFoundException{
+        if (userRepo.findAll() == null){
+            throw new DataNotFoundException("Нет пользователей в базе данных...");
+        }
+        return (List<UserEntity>) userRepo.findAll();
+    }
+
+    public Integer deleteUser(Integer id) throws DataNotFoundException {
+        if (userRepo.findById(id) == null) {
+            throw new DataNotFoundException("Нет пользователя с таким id!");
+        }
         userRepo.deleteById(id);
         return id;
     }
 
-    public Iterable<UserEntity> showAll(){
-        return userRepo.findAll();
-    }
 }
